requirements.txt
pydantic
sqlalchemy
pytest
typer
python-dotenv
google-generativeai
mypy
pytest-cov
typing-extensions